Move jobcreator to jobtype
